requirements.txt
-c "https://raw.githubusercontent.com/apache/airflow/constraints-2.10.0/constraints-3.12.txt"
yfinance==0.2.66
pandas==2.1.4
ta==0.11.0
plotly==6.3.1
apache-airflow==2.10.0
